+++ src/index.ts
@@ -1,13 +1,9 @@
-/// <reference path="../types/is.d.ts" />
-
-import * as isLib from 'is_js';
-const is: Is = isLib;
-
+import { tst, TSType } from './ts_type';
 
 type t_conf = Object;
 type t_conf_object = any;
-type is_fn = (v: any) => boolean;
-type is_fn_custom = (v: any, is: Is, configuration: Configula) => boolean;
+type type_fn = (v: any) => boolean;
+type type_fn_custom = (v: any, tst: TSType, configuration: Configula) => boolean;
 
 export interface IConfigula {
   /**
@@ -54,12 +50,12 @@ export interface IConfigula {
    * @param path a query in to the configuration (. to separate objects, [*] to define arrays)
    * @param type
    */
-  define(path: string, type: any | string | is_fn_custom): this;
+  define(path: string, type: any | string | type_fn_custom): this;
 }
 
 /**
- * Represents a single configuration where every stored key must have a defined type. Type determination is decided by is.js or a custom function which is provided the is.js object.
- * @see is_js
+ * Represents a single configuration where every stored key must have a defined type.
+ *
  */
 export class Configula implements IConfigula {
 
@@ -67,6 +63,10 @@ export class Configula implements IConfigula {
   protected _type: t_conf_object = {};
 
   private __issues: string[] = [];
+
+  constructor(private __of: any | TSType = tst) {
+
+  }
 
   clear() {
     this._type = {};
@@ -123,7 +123,7 @@ export class Configula implements IConfigula {
     return this.__traverse(path, value, `write`);
   }
 
-  define(path: string, type: any | string | is_fn_custom) {
+  define(path: string, type: any | string | type_fn_custom) {
     return this.__traverse(path, type, `define`);
   }
 
@@ -139,7 +139,7 @@ export class Configula implements IConfigula {
 
     if (isNestingObject) {
       for (const k in v) {
-        const _path = (!path) ? `${k}` : ((is.array(v)) ? `${path}[${k}]` : `${path}.${k}`);
+        const _path = (!path) ? `${k}` : ((Array.isArray(v)) ? `${path}[${k}]` : `${path}.${k}`);
 
         this.__traverse(_path, v[k], method);
       }
@@ -161,7 +161,7 @@ export class Configula implements IConfigula {
     const typeInflectionPoint = typeSteps.pop() as string;
     const typeParent = this._getPath(this._type, typeSteps);
 
-    if (is.undefined(value)) {
+    if (tst.undefined(value)) {
       this.__issues.push(`${typeSteps.join('.')} will not accept undefined as a value.`);
       return this;
     }
@@ -198,7 +198,7 @@ export class Configula implements IConfigula {
   /**
    * Write a type definition in to an leaf-path
    */
-  protected _define(path: string, type: any | string | is_fn_custom, issues: string[]): this {
+  protected _define(path: string, type: any | string | type_fn_custom, issues: string[]): this {
     const steps = this._steps(this.__parsePathGlobIndex(path));
     const inflectionPoint = steps.pop() as string;
     const typeParent = this._createPath(this._type, steps);
@@ -208,17 +208,29 @@ export class Configula implements IConfigula {
   }
 
   /**
+   * there are four conditions which could match:
+   *
+   * 1. a literal value matches with `===`
+   * 2. a custom function
+   * 3. a regular expression matches against a string (value will be coerced)
+   * 4. a type matches from `of-ts`
    *
    * @param _type a type function or is.js method
    * @param value the value to type-check
    * @return true if type-check permits the value
    */
-  protected _isPermitted(_type: is_fn_custom | string, value: any): boolean {
-    if (typeof _type === 'function') {
-      return _type(value, is, this);
+  protected _isPermitted(_type: type_fn_custom | any | RegExp, value: any): boolean {
+    if (_type === value) {
+      return true;
     }
-    else if (typeof _type === 'string') {
-      return ((is as any)[_type] as is_fn).call(is, value);
+    else if (tst.function(_type)) {
+      return _type(value, this.__of, this);
+    }
+    else if (tst.string(value) && tst.regExp(_type)) {
+      return (value + '').match(_type) !== null;
+    }
+    else if (tst.string(_type) && this.__of[_type]) {
+      return ((this.__of as any)[_type] as type_fn).call(this.__of, value);
     }
     else {
       return false;
@@ -334,6 +346,7 @@ export class Configula implements IConfigula {
   }
 
   /**
+   * is array or pure object (not built from new but {})
    * @param o is value an object or a nested object which defines the path further
    *
    * ```
@@ -341,7 +354,7 @@ export class Configula implements IConfigula {
    * ```
    */
   private __isNestingObject(o: any): boolean {
-    return is.array(o) || (is.object(o) && o.constructor.name === `Object`);
+    return tst.array(o) || tst.object(o);
   }
 
 }